tradingbotpaper/requirements.txt
ccxt
pandas
numpy
pyyaml
rich